+++ src-input/duk_js_compiler.c
@@ -7824,7 +7824,7 @@ DUK_INTERNAL void duk_js_compile(duk_hthread *thr, const duk_uint8_t *src_buffer
 
 	if (safe_rc != DUK_EXEC_SUCCESS) {
 		DUK_D(DUK_DPRINT("compilation failed: %!T", duk_get_tval(ctx, -1)));
-		duk_throw(ctx);
+		(void) duk_throw(ctx);
 	}
 
 	/* [ ... template ] */