+++ tests/api/test-throw.c
@@ -1,9 +1,11 @@
 /*===
-*** test_1 (duk_safe_call)
+*** test_basic (duk_safe_call)
 ==> rc=1, result='throw me'
+*** test_return (duk_safe_call)
+==> rc=1, result='throw me too'
 ===*/
 
-static duk_ret_t test_1(duk_context *ctx, void *udata) {
+static duk_ret_t test_basic(duk_context *ctx, void *udata) {
 	(void) udata;
 
 	duk_push_string(ctx, "throw me");
@@ -11,6 +13,14 @@ static duk_ret_t test_1(duk_context *ctx, void *udata) {
 	return 0;
 }
 
+static duk_ret_t test_return(duk_context *ctx, void *udata) {
+	(void) udata;
+
+	duk_push_string(ctx, "throw me too");
+	return duk_throw(ctx);
+}
+
 void test(duk_context *ctx) {
-	TEST_SAFE_CALL(test_1);
+	TEST_SAFE_CALL(test_basic);
+	TEST_SAFE_CALL(test_return);
 }